requirements.txt
APScheduler==3.6.3
astroid==2.9.3
cachetools==4.2.2
certifi==2021.10.8
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pylint==2.12.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
wrapt==1.13.3
